Committing code so far. It has fundamental issues, so gonna rewrite.
In the attempt to shoe-horn ModelId everywhere, the code became unbearably messy.
And while doing so, I realized a fundamental flaw that allowed z3 to keep gaming us.
We started using separate symbols for different models, but within each model,
we used the same symbols for Pre and Post states. This was problematic with Implication counter examples.
If we want to specify the synthesis constraints for an implication counter example,
then we have a couple of options:

(1) (pre_state_desc & invar_expr) => (post_state_desc & invar_expr)
(2) (pre_state_desc => ~invar_expr) OR ((pre_state_desc => invar_expr) & (post_state_desc => invar_expr))

Both approaches use implications, and whenever z3 sees implications, the little piece of crap smiles like a devil.

Also, the first one is problematic, even if z3 wasn't being evil with implications. If it finds a model
where pre_state_desc is true and invar_expr is true, it'll be forced to prove post_state_desc to be true
in the same model! Which won't work, because pre_state != post_state in general.

So, the only real solution to getting rid of implications is to have separate symbols for pre and post states.

Signed-off-by: Parth Thakkar <thakkarparth007@gmail.com>