declare-funs to define-funs, and witness ideas
The Witness idea implementation uses foralls still,
which can be expanded in python. However, not doing that yet.

It seems that z3 spends a lot of time in just simplifying
the constraints. The test_synth.sy file got to 650KB at one
point, and then Madhu ran check-simplify on it, and found
the generated file to become 255MB. That's insane, and
shouldn't happen.

Now, I'm gonna try the new ideas from Madhu's emails

Signed-off-by: Parth Thakkar <thakkarparth007@gmail.com>